+++ CHANGELOG.md
@@ -1,3 +1,9 @@
-## 0.0.1
+## [0.0.2] - 2023-4-17
+
+* Fixing line animation
+* Fixing on size of chart
+* Add more parameter to customize chart
+
+## [0.0.1] - 2023-4-16
 
 * Very initial release, to include the library in project